+++ anybox/recipe/openerp/vcs/__init__.py
@@ -1,3 +1,4 @@
+from zc.buildout import UserError
 from .base import UpdateError  # noqa
 from .hg import HgRepo
 from .bzr import BzrBranch
@@ -12,12 +13,16 @@ SUPPORTED['git'] = GitRepo
 SUPPORTED['svn'] = SvnCheckout
 
 
-def get_update(vcs_type, target_dir, url, revision, **options):
-    """General entry point."""
+def repo(vcs_type, target_dir, url, **options):
     cls = SUPPORTED.get(vcs_type)
     if cls is None:
-        raise ValueError("Unsupported VCS type: %r" % vcs_type)
+        raise UserError("Unsupported VCS type: %r" % vcs_type)
 
     # case of standalon addon (see launchpad #1012899)
     target_dir = cls.fix_target(target_dir)
-    cls(target_dir, url, **options)(revision)
+    return cls(target_dir, url, **options)
+
+
+def get_update(vcs_type, target_dir, url, revision, **options):
+    """General entry point."""
+    return repo(vcs_type, target_dir, url, **options)(revision)