+++ CHANGES.rst
@@ -7,6 +7,8 @@ with any higher released unstable versions.
 
 1.9.0 (unreleased)
 ------------------
+- lanchpad #1294020: vcs-revert=on-merge option, and implementation
+  for bzr
 - launchpad #1274174: new feature for automatic merge of remote VCS sources
   (bzr only for now)
 - launchpad #1281262: new ``openerp-log-level`` modifier in ``openerp_scripts``